+++ src/main/java/com/lloll/myro/domain/eventLog/api/EventLogController.java
@@ -1,0 +1,4 @@
+package com.lloll.myro.domain.eventLog.api;
+
+public class EventLogController {
+}
