+++ src/main/java/com/lloll/myro/domain/notification/api/NotificationController.java
@@ -1,0 +1,4 @@
+package com.lloll.myro.domain.notification.api;
+
+public class NotificationController {
+}
