+++ src/main/java/com/lloll/myro/domain/schedule/application/ScheduleService.java
@@ -1,0 +1,4 @@
+package com.lloll.myro.domain.schedule.application;
+
+public class ScheduleService {
+}
