+++ src/main/java/com/lloll/myro/domain/user/api/UserController.java
@@ -1,4 +1,4 @@
-package com.lloll.myro.user.controller;
+package com.lloll.myro.domain.user.api;
 
 import org.springframework.web.bind.annotation.RestController;
