--- src/main/java/com/lloll/myro/eventLog/controller/EventLogController.java
@@ -1,4 +1,0 @@
-package com.lloll.myro.eventLog.controller;
-
-public class EventLogController {
-}
