--- src/main/java/com/lloll/myro/notification/controller/NotificationController.java
@@ -1,4 +1,0 @@
-package com.lloll.myro.notification.controller;
-
-public class NotificationController {
-}
