--- src/main/java/com/lloll/myro/schedule/api/ScheduleContoller.java
@@ -1,4 +1,0 @@
-package com.lloll.myro.schedule.api;
-
-public class ScheduleContoller {
-}
